Initialize Product and Variant

+++ src/main/java/dev/realtards/kuenyawz/services/AccountServiceImpl.java
@@ -1,14 +1,16 @@
 package dev.realtards.kuenyawz.services;
 
 import dev.realtards.kuenyawz.dtos.account.*;
+import dev.realtards.kuenyawz.entities.Account;
 import dev.realtards.kuenyawz.exceptions.AccountExistsException;
 import dev.realtards.kuenyawz.exceptions.AccountNotFoundException;
 import dev.realtards.kuenyawz.exceptions.InvalidPasswordException;
 import dev.realtards.kuenyawz.exceptions.PasswordMismatchException;
-import dev.realtards.kuenyawz.models.Account;
+import dev.realtards.kuenyawz.mapper.AccountMapper;
 import dev.realtards.kuenyawz.repositories.AccountRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.annotation.Primary;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -16,16 +18,21 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@Primary
 @RequiredArgsConstructor
 @Slf4j
 public class AccountServiceImpl implements AccountService {
 
 	private final AccountRepository accountRepository;
 	private final PasswordEncoder passwordEncoder;
+	private final AccountMapper accountMapper;
 
 	@Override
-	public List<Account> getAllAccounts() {
-		return accountRepository.findAll();
+	public List<AccountSecureDto> getAllAccounts() {
+		return accountRepository.findAll()
+			.stream()
+			.map(accountMapper::fromEntity)
+			.toList();
 	}
 
 	@Override
@@ -49,19 +56,24 @@ public class AccountServiceImpl implements AccountService {
 
 	@Override
 	public Account getAccount(long accountId) {
-		Optional<Account> result = accountRepository.findById(accountId);
-		if (result.isEmpty()) {
-			throw new AccountNotFoundException();
-		}
-		log.info("RETRIEVED: {}", result.get());
+		Account account = accountRepository.findById(accountId)
+			.orElseThrow(AccountNotFoundException::new);
+		log.info("RETRIEVED by ID: {}", account);
+		return account;
+	}
 
-		return result.get();
+	@Override
+	public Account getAccount(String email) {
+		Account account = accountRepository.findByEmail(email)
+			.orElseThrow(AccountNotFoundException::new);
+		log.info("RETRIEVED by EMAIL: {}", account);
+		return account;
 	}
 
 	@Override
 	public Account updateAccount(Long accountId, AccountPutDto account) {
 		Account existingAccount = accountRepository.findById(accountId)
-			.orElseThrow(AccountNotFoundException::new);
+			.orElseThrow(() -> new AccountNotFoundException("Account with ID '" + accountId + "' not found"));
 
 		existingAccount.setFullName(account.getFullName());
 		existingAccount.setEmail(account.getEmail());
@@ -75,6 +87,10 @@ public class AccountServiceImpl implements AccountService {
 
 	@Override
 	public void deleteAccount(long accountId) {
+		if (!accountRepository.existsById(accountId)) {
+			throw new AccountNotFoundException("Account with ID '" + accountId + "' not found");
+		}
+
 		accountRepository.deleteById(accountId);
 		log.info("DELETED: {}", accountId);
 	}
@@ -82,7 +98,7 @@ public class AccountServiceImpl implements AccountService {
 	@Override
 	public Account patchAccount(Long accountId, AccountPatchDto accountPatchDto) {
 		final Account existingAccount = accountRepository.findById(accountId)
-			.orElseThrow(AccountNotFoundException::new);
+			.orElseThrow(() -> new AccountNotFoundException("Account with ID '" + accountId + "' not found"));
 
 		Optional.ofNullable(accountPatchDto.getFullName())
 			.ifPresent(existingAccount::setFullName);
@@ -105,7 +121,7 @@ public class AccountServiceImpl implements AccountService {
 	@Override
 	public Account updatePassword(Long accountId, PasswordUpdateDto passwordUpdateDto) {
 		Account existingAccount = accountRepository.findById(accountId)
-			.orElseThrow(AccountNotFoundException::new);
+			.orElseThrow(() -> new AccountNotFoundException("Account with ID '" + accountId + "' not found"));
 
 		// checks whether the current password matches
 		log.debug("DTO Current password: {}", passwordUpdateDto.getCurrentPassword());
@@ -131,7 +147,7 @@ public class AccountServiceImpl implements AccountService {
 	@Override
 	public Account updatePrivilege(Long accountId, PrivilegeUpdateDto privilegeUpdateDto) {
 		Account existingAccount = accountRepository.findById(accountId)
-			.orElseThrow(AccountNotFoundException::new);
+			.orElseThrow(() -> new AccountNotFoundException("Account with ID '" + accountId + "' not found"));
 
 		existingAccount.setPrivilege(privilegeUpdateDto.getPrivilege());
 		existingAccount = accountRepository.save(existingAccount);
